opendingux: Link etc/TZ to etc/local/TZ
This allows the timezone to be set by the user. If we do not disable the
timezone option in the buildroot configuration, buildroot will replace the link
with its own link pointing to a UTC timzeone file which we do not want. But
since UTC is the default anyway it doesnt matter if we just remove it.

Signed-off-by: Jens Nyberg <jens.nyberg@gmail.com>